Move fam views to their own module.
